requirements.txt
django==4.1.11
djangorestframework==3.14.0
django-filter==2.4.0

requests==2.31

pytest==7.2.0
pytest-django==4.5.2
model-bakery==1.9.0

Pillow==9.3.0